pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smriti-memcore"
version = "1.0.1"
description = "A neuro-inspired memory architecture for AI agents — combines a Semantic Palace graph, capacity-bounded Working Memory, and asynchronous consolidation."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Shivam Tyagi", email = "shivamtyagi18@gmail.com"},
]
keywords = [
    "llm", "memory", "ai-agents", "cognitive-architecture",
    "knowledge-graph", "rag", "long-term-memory",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.28.0",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
faiss = [
    "faiss-cpu>=1.7.0",
]
langchain = [
    "langchain-core>=0.3.0",
]
benchmarks = [
    "matplotlib>=3.7.0",
]
mcp = [
    "mcp>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/shivamtyagi18/smriti-memcore"
Repository = "https://github.com/shivamtyagi18/smriti-memcore"
Issues = "https://github.com/shivamtyagi18/smriti-memcore/issues"
Paper = "https://doi.org/10.13140/RG.2.2.25477.82407"

[tool.setuptools.packages.find]
include = ["smriti*"]
exclude = ["tests*", "benchmarks*", "baselines*", "paper*", "results*", "data*", "venv*"]
